styles

+++ src/main/java/com/github/gzougianos/packagraph/graphviz/GraphvizAdapter.java
@@ -1,6 +1,7 @@
 package com.github.gzougianos.packagraph.graphviz;
 
 import com.github.gzougianos.packagraph.GraphLibrary;
+import com.github.gzougianos.packagraph.NodeStyle;
 import com.github.gzougianos.packagraph.PackageNode;
 import com.github.gzougianos.packagraph.PackagraphOptions;
 import com.github.gzougianos.packagraph.analysis.Package;
@@ -13,6 +14,7 @@ import lombok.extern.slf4j.Slf4j;
 
 import java.io.IOException;
 import java.util.*;
+import java.util.function.Supplier;
 import java.util.stream.Collectors;
 
 @Slf4j
@@ -38,6 +40,10 @@ class GraphvizAdapter implements GraphLibrary {
         }
 
         try {
+            if (fileExistsAndCantOverwrite(options))
+                throw new IllegalArgumentException("Output file already exists: " + options.outputFile().getAbsolutePath());
+
+
             Graphviz.fromGraph(g)
                     .render(getFormatOutput(options))
                     .toFile(options.outputFile());
@@ -46,6 +52,10 @@ class GraphvizAdapter implements GraphLibrary {
         } catch (IOException e) {
             log.error("Failed to generate graph image", e);
         }
+    }
+
+    private static boolean fileExistsAndCantOverwrite(PackagraphOptions options) {
+        return options.outputFile().exists() && !options.allowsOverwriteImageOutput();
     }
 
     private static Format getFormatOutput(PackagraphOptions options) {
@@ -76,9 +86,41 @@ class GraphvizAdapter implements GraphLibrary {
 
         return allPackages.stream()
                 .map(packag -> {
-                    Node node = Factory.node(packag.name());
+                    Node node = createNode(packag, options);
                     return new AbstractMap.SimpleEntry<>(packag, node);
                 })
                 .collect(Collectors.toMap(Map.Entry::getKey, Map.Entry::getValue));
+    }
+
+    private Node createNode(Package packag, PackagraphOptions options) {
+        var style = options.styleOf(packag);
+        var node = Factory.node(packag.name());
+        node = applyAttributeIfNotNull(style::shape, "shape", node);
+        node = applyAttributeIfNotNull(style::style, "style", node);
+        node = applyAttributeIfNotNull(style::fillcolor, "fillcolor", node);
+        node = applyAttributeIfNotNull(style::color, "color", node);
+        node = applyAttributeIfNotNull(style::fontcolor, "fontcolor", node);
+        node = applyAttributeIfNotNull(style::fontsize, "fontsize", node);
+        node = applyAttributeIfNotNull(style::fontname, "fontname", node);
+        node = applyAttributeIfNotNull(style::width, "width", node);
+        node = applyAttributeIfNotNull(style::height, "height", node);
+        node = applyAttributeIfNotNull(style::fixedsize, "fixedsize", node);
+        node = applyAttributeIfNotNull(style::tooltip, "tooltip", node);
+        node = applyAttributeIfNotNull(style::url, "url", node);
+        node = applyAttributeIfNotNull(style::target, "target", node);
+        node = applyAttributeIfNotNull(style::layer, "layer", node);
+        node = applyAttributeIfNotNull(style::group, "group", node);
+        node = applyAttributeIfNotNull(style::rank, "rank", node);
+        node = applyAttributeIfNotNull(style::sides, "sides", node);
+        node = applyAttributeIfNotNull(style::peripheries, "peripheries", node);
+        return node;
+    }
+
+    private static Node applyAttributeIfNotNull(Supplier<Object> valueGetter, String attrName, Node node) {
+        var value = valueGetter.get();
+        if (value != null) {
+            return node.with(attrName, value);
+        }
+        return node;
     }
 }
